generator/requirements.txt
accelerate==1.6.0
aiokafka==0.12.0
annotated-types==0.7.0
async-timeout==5.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
diffusers==0.33.1
filelock==3.18.0
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
packaging==25.0
pillow==11.2.1
psutil==7.0.0
pydantic==2.11.3
pydantic-settings==2.9.1
protobuf>=4.25.1
sentencepiece>=0.1.99
pydantic_core==2.33.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sympy==1.13.3
tokenizers==0.21.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
zipp==3.21.0
